// config/release-api/api/api.ts
export interface AddApp {
	/**
	 *
	 * @type {Array<number>}
	 * @memberof AddApp
	 */
	binary: Array<number>;
	/**
	 *
	 * @type {AppMetadata}
	 * @memberof AddApp
	 */
	metadata: AppMetadata;
}

/**
 *
 * @export
 * @enum {string}
 */

export const AppArch = {
	Win32: "Win32",
	Win64: "Win64",
	Linux64: "Linux64",
	Linux32: "Linux32",
} as const;

export type AppArch = typeof AppArch[keyof typeof AppArch];

/**
 *
 * @export
 * @interface AppMetadata
 */
export interface AppMetadata {
	/**
	 *
	 * @type {string}
	 * @memberof AppMetadata
	 */
	name: string;
	/**
	 *
	 * @type {AppVersion}
	 * @memberof AppMetadata
	 */
	version: AppVersion;
	/**
	 *
	 * @type {AppArch}
	 * @memberof AppMetadata
	 */
	arch: AppArch;
}

/**
 *
 * @export
 * @interface AppVersion
 */
export interface AppVersion {
	/**
	 *
	 * @type {number}
	 * @memberof AppVersion
	 */
	minor: number;
	/**
	 *
	 * @type {number}
	 * @memberof AppVersion
	 */
	major: number;
	/**
	 *
	 * @type {number}
	 * @memberof AppVersion
	 */
	revision: number;
}
// ...
export class AppsApi extends BaseAPI {
	/**
	 *
	 * @param {string} name
	 * @param {string} version
	 * @param {AppArch} arch
	 * @param {*} [options] Override http request option.
	 * @throws {RequiredError}
	 * @memberof AppsApi
	 */
	public _delete(name: string, version: string, arch: AppArch, options?: AxiosRequestConfig) {
		return AppsApiFp(this.configuration)
			._delete(name, version, arch, options)
			.then(request => request(this.axios, this.basePath));
	}

	/**
	 *
	 * @param {AddApp} [addApp]
	 * @param {*} [options] Override http request option.
	 * @throws {RequiredError}
	 * @memberof AppsApi
	 */
	public add(addApp?: AddApp, options?: AxiosRequestConfig) {
		return AppsApiFp(this.configuration)
			.add(addApp, options)
			.then(request => request(this.axios, this.basePath));
	}

	/**
	 *
	 * @param {string} name
	 * @param {*} [options] Override http request option.
	 * @throws {RequiredError}
	 * @memberof AppsApi
	 */
	public getAllMetadata(name: string, options?: AxiosRequestConfig) {
		return AppsApiFp(this.configuration)
			.getAllMetadata(name, options)
			.then(request => request(this.axios, this.basePath));
	}

	/**
	 *
	 * @param {*} [options] Override http request option.
	 * @throws {RequiredError}
	 * @memberof AppsApi
	 */
	public getApps(options?: AxiosRequestConfig) {
		return AppsApiFp(this.configuration)
			.getApps(options)
			.then(request => request(this.axios, this.basePath));
	}

	/**
	 *
	 * @param {string} name
	 * @param {string} version
	 * @param {AppArch} arch
	 * @param {*} [options] Override http request option.
	 * @throws {RequiredError}
	 * @memberof AppsApi
	 */
	public getBinary(name: string, version: string, arch: AppArch, options?: AxiosRequestConfig) {
		return AppsApiFp(this.configuration)
			.getBinary(name, version, arch, options)
			.then(request => request(this.axios, this.basePath));
	}

	/**
	 *
	 * @param {string} name
	 * @param {*} [options] Override http request option.
	 * @throws {RequiredError}
	 * @memberof AppsApi
	 */
	public getLatestVersion(name: string, options?: AxiosRequestConfig) {
		return AppsApiFp(this.configuration)
			.getLatestVersion(name, options)
			.then(request => request(this.axios, this.basePath));
	}

	/**
	 *
	 * @param {string} name
	 * @param {AppArch} arch
	 * @param {*} [options] Override http request option.
	 * @throws {RequiredError}
	 * @memberof AppsApi
	 */
	public getLatestVersion_1(name: string, arch: AppArch, options?: AxiosRequestConfig) {
		return AppsApiFp(this.configuration)
			.getLatestVersion_1(name, arch, options)
			.then(request => request(this.axios, this.basePath));
	}
}